+++ src/main/java/com/example/dodam/domain/diary/DiaryImage.java
@@ -2,7 +2,6 @@ package com.example.dodam.domain.diary;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.Date;
 
@@ -10,8 +9,8 @@ import java.util.Date;
 @Getter
 @Setter
 public class DiaryImage {
-    public Integer id;
-    public Integer userId;
+    public Long id;
+    public Long userId;
     public Date date;
     public String title;
     public String base64Img;
